requirements.txt
ram @ git+https://github.com/xinyu1205/recognize-anything.git@88c2b0ca13e38cca6655f83ad0185271167dbcbf
beautifulsoup4==4.12.3
huggingface==0.0.1
huggingface-hub==0.26.0
lxml==5.3.0
pillow==11.0.0
pillow_heif==0.20.0
rawpy>=0.23.1
pyinstaller==6.12.0
tkinter-tooltip==3.1.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu

--extra-index-url https://download.pytorch.org/whl/cpu
torchaudio==2.6.0+cpu

--extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.21.0+cpu

